add GitHub action workflow to repo
With this, we will get a GitHub action that will run every time new
pushes are made and pull requests opened.

The GitHub action workflow verifies we are still able to build a new
cluster and to run the HelloWorld workflow in the newly-built cluster.

This should give us more confidence that changes made to the GitHub repo
at least provide us with a working cluster.

The GitHub action also includes a final step that renders the output of
the HelloWorld workflow as a piece of ASCII art, also useful for
checking whether we are still generating valid TIFF outputs.