SpaceEye - Adjusting reading windows for chips that fall close to the raster's boarder  (#183)
In the SpaceEye workflow, if the input geometry yields a window that is
smaller than the chip size, we adjust it by splitting the difference
between the chip's and window's width (height) half way to the
right/left (top/bottom) sides of the original window. However if the
input geometry falls close to the border of the raster, we might not be
able to completely increase the window in one of the sides and end up
with a window smaller than the chip size. Because of that, the workflow
fails with "_dim_size cannot be smaller than chip_size. Please consider
reducing the step/chip size or increasing the input geometry._"

This PR updates the `_adjust_roi_window` method, so the window is
adjusted from both sides until reaching the chip size or reaching the
border of the raster.

This fixes #180.