Updating url for bitnami charts repo, redis and rabbitmq image tags  (#207)
This PR fixes an issue in which cluster setup would fail with an `Error:
could not download chart: invalid_reference: invalid tag` error message
when installing rabbitmq. This is related to [this issue in bitnami
repo](https://github.com/bitnami/charts/issues/30582), and fixes issue
#203.

This PR also updates the redis and rabbitmq image tags to the latest
versions to ensure compatibility with the latest dependencies and
security patches.